refactor: delete all basic types from "cqp-core"
all those types are moved to "cqp-api"

The "cqp-core" declaring dependency "cqp-api" as transitive
