--- cqp-core/src/main/java/com/quantori/cqp/core/model/Criteria.java
@@ -1,4 +1,0 @@
-package com.quantori.cqp.core.model;
-
-/** A criteria interface that is used for filtering molecules by their properties. */
-public interface Criteria {}
