--- cqp-core/src/main/java/com/quantori/cqp/core/model/ExactParams.java
@@ -1,15 +1,0 @@
-package com.quantori.cqp.core.model;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import lombok.experimental.SuperBuilder;
-
-/** An object to store parameters of exact search for molecules and reactions. */
-@SuperBuilder
-@Getter
-@Setter
-@ToString(callSuper = true)
-@EqualsAndHashCode(callSuper = true)
-public class ExactParams extends SearchParams {}
