--- cqp-core/src/main/java/com/quantori/cqp/core/model/SearchType.java
@@ -1,8 +1,0 @@
-package com.quantori.cqp.core.model;
-
-public enum SearchType {
-  exact,
-  substructure,
-  similarity,
-  all
-}
