--- cqp-core/src/main/java/com/quantori/cqp/core/model/StorageUploadItem.java
@@ -1,7 +1,0 @@
-package com.quantori.cqp.core.model;
-
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-
-@JsonTypeInfo(use = JsonTypeInfo.Id.CLASS)
-public interface StorageUploadItem { // U - storage Upload item
-}
