+++ cqp-core/src/main/java/com/quantori/cqp/core/source/CqpService.java
@@ -5,18 +5,18 @@ import akka.actor.typed.ActorSystem;
 import akka.actor.typed.javadsl.AskPattern;
 import akka.actor.typed.receptionist.Receptionist;
 import akka.actor.typed.receptionist.ServiceKey;
+import com.quantori.cqp.api.DataStorage;
+import com.quantori.cqp.api.SearchItem;
+import com.quantori.cqp.api.StorageItem;
+import com.quantori.cqp.api.StorageUploadItem;
+import com.quantori.cqp.api.model.StorageRequest;
 import com.quantori.cqp.core.model.DataSource;
-import com.quantori.cqp.core.model.DataStorage;
 import com.quantori.cqp.core.model.DataUploadItem;
 import com.quantori.cqp.core.model.ErrorType;
 import com.quantori.cqp.core.model.PipelineStatistics;
 import com.quantori.cqp.core.model.SearchError;
-import com.quantori.cqp.core.model.SearchItem;
 import com.quantori.cqp.core.model.SearchRequest;
 import com.quantori.cqp.core.model.SearchResult;
-import com.quantori.cqp.core.model.StorageItem;
-import com.quantori.cqp.core.model.StorageRequest;
-import com.quantori.cqp.core.model.StorageUploadItem;
 import com.quantori.cqp.core.model.TransformationStep;
 import java.time.Duration;
 import java.util.List;
