+++ cqp-core/src/main/java/com/quantori/cqp/core/source/DataSourceActor.java
@@ -21,12 +21,12 @@ import akka.stream.javadsl.Keep;
 import akka.stream.javadsl.Merge;
 import akka.stream.javadsl.Sink;
 import akka.stream.javadsl.Source;
+import com.quantori.cqp.api.SearchItem;
+import com.quantori.cqp.api.SearchIterator;
+import com.quantori.cqp.api.StorageItem;
 import com.quantori.cqp.core.model.ErrorType;
 import com.quantori.cqp.core.model.SearchError;
-import com.quantori.cqp.core.model.SearchItem;
-import com.quantori.cqp.core.model.SearchIterator;
 import com.quantori.cqp.core.model.SearchRequest;
-import com.quantori.cqp.core.model.StorageItem;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
