+++ cqp-core/src/main/java/com/quantori/cqp/core/source/Loader.java
@@ -10,10 +10,10 @@ import akka.stream.Supervision;
 import akka.stream.javadsl.Keep;
 import akka.stream.javadsl.Sink;
 import akka.stream.javadsl.Source;
+import com.quantori.cqp.api.StorageUploadItem;
 import com.quantori.cqp.core.model.DataSource;
 import com.quantori.cqp.core.model.DataUploadItem;
 import com.quantori.cqp.core.model.PipelineStatistics;
-import com.quantori.cqp.core.model.StorageUploadItem;
 import com.quantori.cqp.core.model.TransformationStep;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
