+++ cqp-core/src/main/java/com/quantori/cqp/core/source/SearchActor.java
@@ -10,13 +10,13 @@ import akka.actor.typed.javadsl.ReceiveBuilder;
 import akka.actor.typed.javadsl.TimerScheduler;
 import akka.actor.typed.receptionist.ServiceKey;
 import akka.pattern.StatusReply;
-import com.quantori.cqp.core.model.DataStorage;
-import com.quantori.cqp.core.model.SearchItem;
-import com.quantori.cqp.core.model.SearchIterator;
+import com.quantori.cqp.api.DataStorage;
+import com.quantori.cqp.api.SearchItem;
+import com.quantori.cqp.api.SearchIterator;
+import com.quantori.cqp.api.StorageItem;
+import com.quantori.cqp.api.model.StorageRequest;
 import com.quantori.cqp.core.model.SearchRequest;
 import com.quantori.cqp.core.model.SearchResult;
-import com.quantori.cqp.core.model.StorageItem;
-import com.quantori.cqp.core.model.StorageRequest;
 import java.time.Duration;
 import java.util.Collection;
 import java.util.HashMap;
