+++ cqp-core/src/main/java/com/quantori/cqp/core/source/SearchSourceActor.java
@@ -9,8 +9,8 @@ import akka.actor.typed.javadsl.Receive;
 import akka.actor.typed.receptionist.Receptionist;
 import akka.actor.typed.receptionist.ServiceKey;
 import akka.pattern.StatusReply;
-import com.quantori.cqp.core.model.DataStorage;
-import com.quantori.cqp.core.model.StorageItem;
+import com.quantori.cqp.api.DataStorage;
+import com.quantori.cqp.api.StorageItem;
 import java.util.Map;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
