+++ cqp-core/src/main/java/com/quantori/cqp/core/source/Searcher.java
@@ -3,12 +3,12 @@ package com.quantori.cqp.core.source;
 import akka.actor.typed.ActorRef;
 import akka.actor.typed.javadsl.ActorContext;
 import akka.actor.typed.javadsl.AskPattern;
+import com.quantori.cqp.api.SearchItem;
+import com.quantori.cqp.api.SearchIterator;
+import com.quantori.cqp.api.StorageItem;
 import com.quantori.cqp.core.model.FetchWaitMode;
-import com.quantori.cqp.core.model.SearchItem;
-import com.quantori.cqp.core.model.SearchIterator;
 import com.quantori.cqp.core.model.SearchRequest;
 import com.quantori.cqp.core.model.SearchResult;
-import com.quantori.cqp.core.model.StorageItem;
 import java.time.Duration;
 import java.util.List;
 import java.util.Map;
