+++ cqp-core/src/main/java/com/quantori/cqp/core/source/SourceRootActor.java
@@ -11,9 +11,9 @@ import akka.actor.typed.javadsl.ReceiveBuilder;
 import akka.actor.typed.receptionist.Receptionist;
 import akka.actor.typed.receptionist.ServiceKey;
 import akka.pattern.StatusReply;
-import com.quantori.cqp.core.model.DataStorage;
-import com.quantori.cqp.core.model.StorageItem;
-import com.quantori.cqp.core.model.StorageUploadItem;
+import com.quantori.cqp.api.DataStorage;
+import com.quantori.cqp.api.StorageItem;
+import com.quantori.cqp.api.StorageUploadItem;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.HashMap;
