+++ cqp-core/src/main/java/com/quantori/cqp/core/source/UploadSourceActor.java
@@ -9,12 +9,12 @@ import akka.actor.typed.javadsl.ActorContext;
 import akka.actor.typed.javadsl.Behaviors;
 import akka.actor.typed.javadsl.Receive;
 import akka.pattern.StatusReply;
+import com.quantori.cqp.api.DataStorage;
+import com.quantori.cqp.api.ItemWriter;
+import com.quantori.cqp.api.StorageUploadItem;
 import com.quantori.cqp.core.model.DataSource;
-import com.quantori.cqp.core.model.DataStorage;
 import com.quantori.cqp.core.model.DataUploadItem;
-import com.quantori.cqp.core.model.ItemWriter;
 import com.quantori.cqp.core.model.PipelineStatistics;
-import com.quantori.cqp.core.model.StorageUploadItem;
 import com.quantori.cqp.core.model.TransformationStep;
 import java.util.LinkedList;
 import java.util.Queue;
