+++ cqp-core/src/main/resources/akka-cluster.conf
@@ -30,7 +30,7 @@ akka {
         "com.quantori.cqp.core.source.SearchActor$Close" = jackson-cbor
         "com.quantori.cqp.core.model.SearchResult" = jackson-cbor
         "com.quantori.cqp.core.model.SearchRequest" = jackson-cbor
-        "com.quantori.cqp.core.model.StorageRequest" = jackson-cbor
+        "com.quantori.cqp.api.model.StorageRequest" = jackson-cbor
       }
   }
 }
