+++ cqp-core/src/test/java/com/quantori/cqp/core/source/CqpServiceTest.java
@@ -9,15 +9,15 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.when;
 
 import akka.actor.typed.ActorSystem;
+import com.quantori.cqp.api.DataStorage;
+import com.quantori.cqp.api.ItemWriter;
+import com.quantori.cqp.api.SearchIterator;
+import com.quantori.cqp.api.model.StorageRequest;
 import com.quantori.cqp.core.configuration.ClusterConfigurationProperties;
 import com.quantori.cqp.core.configuration.ClusterProvider;
 import com.quantori.cqp.core.model.DataSource;
-import com.quantori.cqp.core.model.DataStorage;
-import com.quantori.cqp.core.model.ItemWriter;
-import com.quantori.cqp.core.model.SearchIterator;
 import com.quantori.cqp.core.model.SearchRequest;
 import com.quantori.cqp.core.model.SearchResult;
-import com.quantori.cqp.core.model.StorageRequest;
 import com.quantori.cqp.core.model.TransformationStep;
 import com.quantori.cqp.core.model.TransformationStepBuilder;
 import java.util.ArrayList;
