+++ cqp-core/src/test/java/com/quantori/cqp/core/source/SearcherBufferTest.java
@@ -3,12 +3,12 @@ package com.quantori.cqp.core.source;
 import akka.actor.testkit.typed.javadsl.ActorTestKit;
 import akka.actor.typed.ActorRef;
 import akka.pattern.StatusReply;
-import com.quantori.cqp.core.model.DataStorage;
-import com.quantori.cqp.core.model.ItemWriter;
-import com.quantori.cqp.core.model.SearchIterator;
+import com.quantori.cqp.api.DataStorage;
+import com.quantori.cqp.api.ItemWriter;
+import com.quantori.cqp.api.SearchIterator;
+import com.quantori.cqp.api.model.StorageRequest;
 import com.quantori.cqp.core.model.SearchRequest;
 import com.quantori.cqp.core.model.SearchResult;
-import com.quantori.cqp.core.model.StorageRequest;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Iterator;
