+++ cqp-core/src/test/java/com/quantori/cqp/core/source/SearcherTest.java
@@ -13,10 +13,10 @@ import akka.actor.typed.javadsl.ActorContext;
 import akka.actor.typed.javadsl.Behaviors;
 import akka.actor.typed.javadsl.Receive;
 import akka.pattern.StatusReply;
-import com.quantori.cqp.core.model.SearchIterator;
+import com.quantori.cqp.api.SearchIterator;
+import com.quantori.cqp.api.model.StorageRequest;
 import com.quantori.cqp.core.model.SearchRequest;
 import com.quantori.cqp.core.model.SearchResult;
-import com.quantori.cqp.core.model.StorageRequest;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Iterator;
