+++ cqp-core/src/test/java/com/quantori/cqp/core/source/TestStorageItem.java
@@ -2,7 +2,7 @@ package com.quantori.cqp.core.source;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.quantori.cqp.core.model.StorageItem;
+import com.quantori.cqp.api.StorageItem;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
